+++ Controller/Withdraw/Waiver/Context.php
@@ -7,6 +7,7 @@ declare(strict_types=1);
 
 namespace MageMe\EUWithdrawal\Controller\Withdraw\Waiver;
 
+use MageMe\EUWithdrawal\Model\Geo\CountryScope;
 use MageMe\EUWithdrawal\Model\Waiver\WaiverTextHasher;
 use MageMe\EUWithdrawal\Model\Waiver\WaiverTextResolver;
 use MageMe\EUWithdrawal\Service\DigitalContentDetector;
@@ -27,6 +28,7 @@ class Context implements HttpGetActionInterface
      * @param WaiverTextHasher $hasher
      * @param StoreManagerInterface $storeManager
      * @param JsonFactory $jsonFactory
+     * @param CountryScope $countryScope
      */
     public function __construct(
         private readonly CheckoutSession $checkoutSession,
@@ -35,6 +37,7 @@ class Context implements HttpGetActionInterface
         private readonly WaiverTextHasher $hasher,
         private readonly StoreManagerInterface $storeManager,
         private readonly JsonFactory $jsonFactory,
+        private readonly CountryScope $countryScope,
     ) {
     }
 
@@ -46,6 +49,13 @@ class Context implements HttpGetActionInterface
     public function execute(): ResultInterface
     {
         $quote = $this->checkoutSession->getQuote();
+        if (!$this->countryScope->quoteInScope($quote)) {
+            return $this->jsonFactory->create()->setData([
+                'items' => [],
+                'locale' => null,
+                'jurisdiction' => null,
+            ]);
+        }
         $digital = $this->detector->filterDigitalItems($quote->getAllVisibleItems());
         $locale = (string) $this->storeManager->getStore()->getConfig('general/locale/code');
         $jurisdiction = \strtoupper(\substr((string) ($quote->getBillingAddress()?->getCountryId() ?? ''), 0, 2));